requirements.txt
agno
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
alembic
httpx
feedparser
beautifulsoup4
requests
celery
redis
apscheduler
dash
plotly
pandas
python-dateutil
pytz